requirements.txt
aiohttp==1.0.5
discord.py==0.16.11